+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * update `Requires at least` in plugin file
+* add VCS checkouts automatically, thanks @Raruto
 
 #### 2.4.1 / 2019-05-09
 * a11y updates for settings tabs